requirements.txt
streamlit>=1.25.0 #Web app framework / Framework para apps web
google-generativeai>=0.0.1 # Google Gemini AI SDK/SDK de IA Gemini do Google
tmdbv3api>=1.6.0 # TMDb API SDK / Biblioteca oficial TMDb
requests>=2.31.0 # HTTP requests / Requesições HTTP
python-dotenv>=1.0.0 
pyautogui>=0.9.54 # GUI automation / Automação de interface gráfica python --version